Convert dedup plugin to TO
